Refactored code with new MVP architecture
