--- app/src/main/java/com/andigeeky/weddinginvitation/SplashScreen.java
@@ -1,40 +1,0 @@
-package com.andigeeky.weddinginvitation;
-
-import android.content.Intent;
-import android.os.Bundle;
-import android.os.Handler;
-import android.support.v7.app.AppCompatActivity;
-
-public class SplashScreen extends AppCompatActivity {
-
-    private Handler mHandler;
-    private Runnable mRunnable;
-    private static final long SPLASH_DURATION = 2500L;
-
-
-    @Override
-    protected void onCreate(Bundle savedInstanceState) {
-        super.onCreate(savedInstanceState);
-
-        mHandler = new Handler();
-        mRunnable = new Runnable() {
-            @Override
-            public void run() {
-                startActivity(new Intent(SplashScreen.this, MainActivity.class));
-                finish();
-            }
-        };
-    }
-
-    @Override
-    protected void onResume() {
-        super.onResume();
-        mHandler.postDelayed(mRunnable, SPLASH_DURATION);
-    }
-
-    @Override
-    protected void onPause() {
-        super.onPause();
-        mHandler.removeCallbacks(mRunnable);
-    }
-}
