--- app/src/main/java/com/andigeeky/weddinginvitation/auth/AccountType.java
@@ -1,11 +1,0 @@
-package com.andigeeky.weddinginvitation.auth;
-
-/**
- * Created by Mamata on 12/15/2017.
- */
-
-public enum AccountType {
-    EMAIL,
-    GOOGLE,
-    FACEBOOK
-}
