--- app/src/main/java/com/andigeeky/weddinginvitation/auth/contract/IRegisterUser.java
@@ -1,8 +1,0 @@
-package com.andigeeky.weddinginvitation.auth.contract;
-
-import com.andigeeky.weddinginvitation.auth.AccountType;
-import com.google.firebase.auth.FirebaseUser;
-
-public interface IRegisterUser {
-    FirebaseUser registerUser(AccountType accountType);
-}
