+++ app/src/main/java/com/andigeeky/weddinginvitation/domain/RemoteRepository.java
@@ -1,6 +1,6 @@
-package com.andigeeky.weddinginvitation.newarchitecture;
+package com.andigeeky.weddinginvitation.domain;
 
-import com.andigeeky.weddinginvitation.auth.vo.User;
+import com.andigeeky.weddinginvitation.model.User;
 
 import io.reactivex.Completable;
