+++ app/src/main/java/com/andigeeky/weddinginvitation/model/User.java
@@ -1,4 +1,4 @@
-package com.andigeeky.weddinginvitation.auth.vo;
+package com.andigeeky.weddinginvitation.model;
 
 
 import java.io.Serializable;
