+++ app/src/main/java/com/andigeeky/weddinginvitation/presentation/UserViewModel.java
@@ -1,8 +1,9 @@
-package com.andigeeky.weddinginvitation.newarchitecture;
+package com.andigeeky.weddinginvitation.presentation;
 
 import android.arch.lifecycle.ViewModel;
 
-import com.andigeeky.weddinginvitation.auth.vo.User;
+import com.andigeeky.weddinginvitation.model.User;
+import com.andigeeky.weddinginvitation.domain.RegisterUseCase;
 
 import io.reactivex.android.schedulers.AndroidSchedulers;
 import io.reactivex.disposables.CompositeDisposable;
